backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database & ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1

# ML Libraries
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.4
xgboost==2.0.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
openpyxl==3.1.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0

# AI Integration (Optional for Week 7-8)
openai==1.3.7

# Docker
docker==6.1.3